+++ evaluator/evaluator_test.go
@@ -26,7 +26,6 @@ import (
 	"github.com/pingcap/tidb/mysql"
 	"github.com/pingcap/tidb/parser"
 	"github.com/pingcap/tidb/parser/opcode"
-	"github.com/pingcap/tidb/sessionctx/varsutil"
 	"github.com/pingcap/tidb/util/charset"
 	"github.com/pingcap/tidb/util/mock"
 	"github.com/pingcap/tidb/util/testleak"
@@ -933,94 +932,6 @@ func (s *testEvaluatorSuite) TestAggFuncAvg(c *C) {
 	expect := types.NewDecFromInt(3)
 	c.Assert(result.Kind(), Equals, types.KindMysqlDecimal)
 	c.Assert(result.GetMysqlDecimal().Compare(expect), Equals, 0)
-}
-
-func (s *testEvaluatorSuite) TestGetTimeValue(c *C) {
-	defer testleak.AfterTest(c)()
-	v, err := GetTimeValue(nil, "2012-12-12 00:00:00", mysql.TypeTimestamp, types.MinFsp)
-	c.Assert(err, IsNil)
-
-	c.Assert(v.Kind(), Equals, types.KindMysqlTime)
-	timeValue := v.GetMysqlTime()
-	c.Assert(timeValue.String(), Equals, "2012-12-12 00:00:00")
-	ctx := mock.NewContext()
-	sessionVars := ctx.GetSessionVars()
-	varsutil.SetSystemVar(sessionVars, "timestamp", types.NewStringDatum(""))
-	v, err = GetTimeValue(ctx, "2012-12-12 00:00:00", mysql.TypeTimestamp, types.MinFsp)
-	c.Assert(err, IsNil)
-
-	c.Assert(v.Kind(), Equals, types.KindMysqlTime)
-	timeValue = v.GetMysqlTime()
-	c.Assert(timeValue.String(), Equals, "2012-12-12 00:00:00")
-
-	varsutil.SetSystemVar(sessionVars, "timestamp", types.NewStringDatum("0"))
-	v, err = GetTimeValue(ctx, "2012-12-12 00:00:00", mysql.TypeTimestamp, types.MinFsp)
-	c.Assert(err, IsNil)
-
-	c.Assert(v.Kind(), Equals, types.KindMysqlTime)
-	timeValue = v.GetMysqlTime()
-	c.Assert(timeValue.String(), Equals, "2012-12-12 00:00:00")
-
-	varsutil.SetSystemVar(sessionVars, "timestamp", types.Datum{})
-	v, err = GetTimeValue(ctx, "2012-12-12 00:00:00", mysql.TypeTimestamp, types.MinFsp)
-	c.Assert(err, IsNil)
-
-	c.Assert(v.Kind(), Equals, types.KindMysqlTime)
-	timeValue = v.GetMysqlTime()
-	c.Assert(timeValue.String(), Equals, "2012-12-12 00:00:00")
-
-	varsutil.SetSystemVar(sessionVars, "timestamp", types.NewStringDatum("1234"))
-
-	tbl := []struct {
-		Expr interface{}
-		Ret  interface{}
-	}{
-		{"2012-12-12 00:00:00", "2012-12-12 00:00:00"},
-		{CurrentTimestamp, time.Unix(1234, 0).Format(types.TimeFormat)},
-		{ZeroTimestamp, "0000-00-00 00:00:00"},
-		{ast.NewValueExpr("2012-12-12 00:00:00"), "2012-12-12 00:00:00"},
-		{ast.NewValueExpr(int64(0)), "0000-00-00 00:00:00"},
-		{ast.NewValueExpr(nil), nil},
-		{&ast.FuncCallExpr{FnName: model.NewCIStr(CurrentTimestamp)}, CurrentTimestamp},
-		{&ast.UnaryOperationExpr{Op: opcode.Minus, V: ast.NewValueExpr(int64(0))}, "0000-00-00 00:00:00"},
-	}
-
-	for i, t := range tbl {
-		comment := Commentf("expr: %d", i)
-		v, err := GetTimeValue(ctx, t.Expr, mysql.TypeTimestamp, types.MinFsp)
-		c.Assert(err, IsNil)
-
-		switch v.Kind() {
-		case types.KindMysqlTime:
-			c.Assert(v.GetMysqlTime().String(), DeepEquals, t.Ret, comment)
-		default:
-			c.Assert(v.GetValue(), DeepEquals, t.Ret, comment)
-		}
-	}
-
-	errTbl := []struct {
-		Expr interface{}
-	}{
-		{"2012-13-12 00:00:00"},
-		{ast.NewValueExpr("2012-13-12 00:00:00")},
-		{ast.NewValueExpr(int64(1))},
-		{&ast.FuncCallExpr{FnName: model.NewCIStr("xxx")}},
-		{&ast.UnaryOperationExpr{Op: opcode.Minus, V: ast.NewValueExpr(int64(1))}},
-	}
-
-	for _, t := range errTbl {
-		_, err := GetTimeValue(ctx, t.Expr, mysql.TypeTimestamp, types.MinFsp)
-		c.Assert(err, NotNil)
-	}
-}
-
-func (s *testEvaluatorSuite) TestIsCurrentTimeExpr(c *C) {
-	defer testleak.AfterTest(c)()
-	v := IsCurrentTimeExpr(ast.NewValueExpr("abc"))
-	c.Assert(v, IsFalse)
-
-	v = IsCurrentTimeExpr(&ast.FuncCallExpr{FnName: model.NewCIStr("CURRENT_TIMESTAMP")})
-	c.Assert(v, IsTrue)
 }
 
 func (s *testEvaluatorSuite) TestEvaluatedFlag(c *C) {
