+++ expression/expression.go
@@ -27,6 +27,9 @@ import (
 	"github.com/pingcap/tidb/util/types"
 )
 
+// EvalAstExpr evaluates ast expression directly.
+var EvalAstExpr func(expr ast.ExprNode, ctx context.Context) (types.Datum, error)
+
 // Expression represents all scalar expression in SQL.
 type Expression interface {
 	fmt.Stringer
