+++ expression/helper.go
@@ -1,4 +1,4 @@
-package evaluator
+package expression
 
 import (
 	"strings"
@@ -78,7 +78,7 @@ func getTimeValue(ctx context.Context, v interface{}, tp byte, fsp int) (d types
 		return d, errors.Trace(errDefaultValue)
 	case *ast.UnaryOperationExpr:
 		// support some expression, like `-1`
-		v, err := Eval(ctx, x)
+		v, err := EvalAstExpr(x, ctx)
 		if err != nil {
 			return d, errors.Trace(err)
 		}
